app/go: reorganize some files
