+++ builtin/app/go/app_test.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/hashicorp/otto/app"
 	"github.com/hashicorp/otto/helper/compile"
-	"github.com/hashicorp/otto/helper/vagrant"
 	"github.com/hashicorp/otto/otto"
 )
 
@@ -72,30 +71,5 @@ func TestApp_importPath(t *testing.T) {
 				Value: "/opt/gopath/src/example.com",
 			},
 		},
-	})
-}
-
-func TestApp_dev(t *testing.T) {
-	otto.Test(t, otto.TestCase{
-		Core: otto.TestCore(t, &otto.TestCoreOpts{
-			Path: filepath.Join("./test-fixtures", "basic", "Appfile"),
-			App:  new(App),
-		}),
-
-		Steps: []otto.TestStep{
-			&vagrant.DevTestStepInit{},
-
-			// Verify we have Go
-			&vagrant.DevTestStepGuestScript{
-				Command: "go version",
-			},
-
-			// Verify we can build immediately (we should be in the directory)
-			&vagrant.DevTestStepGuestScript{
-				Command: "grep '42' <<< $(go build -o test-output && ./test-output)",
-			},
-		},
-
-		Teardown: vagrant.DevTestTeardown,
 	})
 }
